Refactor summary serialization boundary package
Split summary_serialization into focused contracts, findings, plots, and summary assembly modules while keeping the public import path stable through package re-exports.
