+++ apps/server/vibesensor/shared/boundaries/summary_serialization/__init__.py
@@ -1,0 +1,62 @@
+"""Summary payload serialization boundary package.
+
+This package keeps the historical import path stable while splitting the
+serialization boundary into focused modules.
+"""
+
+from ._contracts import (
+    AccelStatisticsLike,
+    AmpVsPhaseRowLike,
+    FreqVsSpeedByFindingSeriesLike,
+    MatchedAmpVsSpeedSeriesLike,
+    PeakTableRowLike,
+    PhaseBoundaryLike,
+    PhaseSegmentLike,
+    PhaseSegmentPlotLike,
+    PhaseSpeedBreakdownRowLike,
+    PlotDataResultLike,
+    SpectrogramResultLike,
+    SpeedBreakdownRowLike,
+)
+from ._findings import annotate_peaks_with_order_labels, serialize_findings
+from ._plots import (
+    serialize_peak_table,
+    serialize_phase_segments,
+    serialize_phase_speed_breakdown,
+    serialize_plot_data,
+    serialize_spectrogram,
+    serialize_speed_breakdown,
+)
+from ._summary import (
+    build_data_quality_dict,
+    build_summary_payload,
+    noise_baseline_db,
+    serialize_origin_summary,
+)
+
+__all__ = [
+    "AccelStatisticsLike",
+    "AmpVsPhaseRowLike",
+    "FreqVsSpeedByFindingSeriesLike",
+    "MatchedAmpVsSpeedSeriesLike",
+    "PeakTableRowLike",
+    "PhaseBoundaryLike",
+    "PhaseSegmentLike",
+    "PhaseSegmentPlotLike",
+    "PhaseSpeedBreakdownRowLike",
+    "PlotDataResultLike",
+    "SpectrogramResultLike",
+    "SpeedBreakdownRowLike",
+    "annotate_peaks_with_order_labels",
+    "build_data_quality_dict",
+    "build_summary_payload",
+    "noise_baseline_db",
+    "serialize_findings",
+    "serialize_origin_summary",
+    "serialize_peak_table",
+    "serialize_phase_segments",
+    "serialize_phase_speed_breakdown",
+    "serialize_plot_data",
+    "serialize_speed_breakdown",
+    "serialize_spectrogram",
+]
