+++ apps/server/vibesensor/shared/boundaries/summary_serialization/_contracts.py
@@ -1,0 +1,189 @@
+"""Structural typing contracts for summary serialization inputs."""
+
+from __future__ import annotations
+
+from collections.abc import Mapping, Sequence
+from typing import Protocol
+
+from vibesensor.domain import DrivingPhase
+
+AccelStatisticsLike = Mapping[str, object]
+
+
+class SpeedBreakdownRowLike(Protocol):
+    @property
+    def speed_range(self) -> str: ...
+    @property
+    def count(self) -> int: ...
+    @property
+    def mean_vibration_strength_db(self) -> float | None: ...
+    @property
+    def max_vibration_strength_db(self) -> float | None: ...
+
+
+class PhaseSpeedBreakdownRowLike(Protocol):
+    @property
+    def phase(self) -> str: ...
+    @property
+    def count(self) -> int: ...
+    @property
+    def mean_speed_kmh(self) -> float | None: ...
+    @property
+    def max_speed_kmh(self) -> float | None: ...
+    @property
+    def mean_vibration_strength_db(self) -> float | None: ...
+    @property
+    def max_vibration_strength_db(self) -> float | None: ...
+
+
+class PeakTableRowLike(Protocol):
+    @property
+    def rank(self) -> int: ...
+    @property
+    def frequency_hz(self) -> float: ...
+    @property
+    def order_label(self) -> str: ...
+    @property
+    def suspected_source(self) -> str: ...
+    @property
+    def max_intensity_db(self) -> float | None: ...
+    @property
+    def median_intensity_db(self) -> float | None: ...
+    @property
+    def p95_intensity_db(self) -> float | None: ...
+    @property
+    def run_noise_baseline_db(self) -> float | None: ...
+    @property
+    def median_vs_run_noise_ratio(self) -> float: ...
+    @property
+    def p95_vs_run_noise_ratio(self) -> float: ...
+    @property
+    def strength_floor_db(self) -> float | None: ...
+    @property
+    def strength_db(self) -> float | None: ...
+    @property
+    def presence_ratio(self) -> float: ...
+    @property
+    def burstiness(self) -> float: ...
+    @property
+    def persistence_score(self) -> float: ...
+    @property
+    def peak_classification(self) -> str: ...
+    @property
+    def typical_speed_band(self) -> str: ...
+
+
+class SpectrogramResultLike(Protocol):
+    @property
+    def x_axis(self) -> str: ...
+    @property
+    def x_label_key(self) -> str: ...
+    @property
+    def x_bins(self) -> Sequence[float]: ...
+    @property
+    def y_bins(self) -> Sequence[float]: ...
+    @property
+    def cells(self) -> Sequence[Sequence[float]]: ...
+    @property
+    def max_amp(self) -> float: ...
+    @property
+    def x_bin_width(self) -> float | None: ...
+    @property
+    def y_bin_width(self) -> float | None: ...
+
+
+class AmpVsPhaseRowLike(Protocol):
+    @property
+    def phase(self) -> str: ...
+    @property
+    def count(self) -> int: ...
+    @property
+    def mean_vib_db(self) -> float: ...
+    @property
+    def max_vib_db(self) -> float | None: ...
+    @property
+    def mean_speed_kmh(self) -> float | None: ...
+
+
+class MatchedAmpVsSpeedSeriesLike(Protocol):
+    @property
+    def label(self) -> str: ...
+    @property
+    def points(self) -> Sequence[tuple[float, float]]: ...
+
+
+class FreqVsSpeedByFindingSeriesLike(Protocol):
+    @property
+    def label(self) -> str: ...
+    @property
+    def matched(self) -> Sequence[tuple[float, float]]: ...
+    @property
+    def predicted(self) -> Sequence[tuple[float, float]]: ...
+
+
+class PhaseSegmentPlotLike(Protocol):
+    @property
+    def phase(self) -> str: ...
+    @property
+    def start_t_s(self) -> float | None: ...
+    @property
+    def end_t_s(self) -> float | None: ...
+
+
+class PhaseBoundaryLike(Protocol):
+    @property
+    def phase(self) -> str: ...
+    @property
+    def t_s(self) -> float | None: ...
+    @property
+    def end_t_s(self) -> float | None: ...
+
+
+class PlotDataResultLike(Protocol):
+    @property
+    def vib_magnitude(self) -> Sequence[tuple[float, float, str]]: ...
+    @property
+    def dominant_freq(self) -> Sequence[tuple[float, float]]: ...
+    @property
+    def amp_vs_speed(self) -> Sequence[tuple[float, float]]: ...
+    @property
+    def amp_vs_phase(self) -> Sequence[AmpVsPhaseRowLike]: ...
+    @property
+    def matched_amp_vs_speed(self) -> Sequence[MatchedAmpVsSpeedSeriesLike]: ...
+    @property
+    def freq_vs_speed_by_finding(self) -> Sequence[FreqVsSpeedByFindingSeriesLike]: ...
+    @property
+    def steady_speed_distribution(self) -> dict[str, float] | None: ...
+    @property
+    def fft_spectrum(self) -> Sequence[tuple[float, float]]: ...
+    @property
+    def fft_spectrum_raw(self) -> Sequence[tuple[float, float]]: ...
+    @property
+    def peaks_spectrogram(self) -> SpectrogramResultLike: ...
+    @property
+    def peaks_spectrogram_raw(self) -> SpectrogramResultLike: ...
+    @property
+    def peaks_table(self) -> Sequence[PeakTableRowLike]: ...
+    @property
+    def phase_segments(self) -> Sequence[PhaseSegmentPlotLike]: ...
+    @property
+    def phase_boundaries(self) -> Sequence[PhaseBoundaryLike]: ...
+
+
+class PhaseSegmentLike(Protocol):
+    @property
+    def phase(self) -> DrivingPhase: ...
+    @property
+    def start_idx(self) -> int: ...
+    @property
+    def end_idx(self) -> int: ...
+    @property
+    def start_t_s(self) -> float: ...
+    @property
+    def end_t_s(self) -> float: ...
+    @property
+    def speed_min_kmh(self) -> float | None: ...
+    @property
+    def speed_max_kmh(self) -> float | None: ...
+    @property
+    def sample_count(self) -> int: ...
